Proyectos de Automatización/Automatización para Principiantes/App de Facturación/requirements.txt
streamlit==1.27.0
streamlit-elements==0.1.0
streamlit-option-menu==0.3.6
pandas==2.0.3
uuid==1.30
python-dotenv==1.0.0
gspread==5.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.31.0